Moved `KeyEntry` to a separate dir.
