+++ yoda/yoda.h
@@ -85,248 +85,9 @@ SOFTWARE.
 #include "policy.h"
 #include "exceptions.h"
 
+#include "api/key_entry/key_entry.h"
+
 namespace yoda {
-
-// TODO(dkorolev): With variadic implementation, the type list for the visitor will be taken from another place.
-template <typename ENTRY>
-struct Container<KeyEntry<ENTRY>> : bricks::metaprogramming::visitor<std::tuple<ENTRY>> {
-  typedef ENTRY T_ENTRY;
-  typedef ENTRY_KEY_TYPE<T_ENTRY> T_KEY;
-
-  T_MAP_TYPE<T_KEY, T_ENTRY> data;
-
-  // The container itself is visitable by entries.
-  // The entries passed in this way are the entries scanned from the Sherlock stream.
-  // The default behavior is to update the contents of the container.
-  virtual void visit(ENTRY& entry) override { data[GetKey(entry)] = entry; }
-};
-
-template <typename ENTRY_BASE_TYPE, typename ENTRY>
-struct StreamListener<ENTRY_BASE_TYPE, std::tuple<KeyEntry<ENTRY>>> {
-  typedef YodaTypes<ENTRY_BASE_TYPE, std::tuple<KeyEntry<ENTRY>>> YT;
-
-  explicit StreamListener(typename YT::T_MQ& mq) : caught_up_(false), entries_seen_(0u), mq_(mq) {}
-
-  struct MQMessageEntry : MQMessage<ENTRY_BASE_TYPE, typename YT::T_SUPPORTED_TYPES_AS_TUPLE> {
-    std::unique_ptr<ENTRY_BASE_TYPE> entry;
-
-    explicit MQMessageEntry(std::unique_ptr<ENTRY_BASE_TYPE>&& entry) : entry(std::move(entry)) {}
-
-    // TODO(dkorolev): Replace this by a variadic implementation. Metaprogramming MapReduce FTW!
-    static_assert(sizeof(is_same_or_compile_error<typename YT::T_VISITABLE_TYPES_AS_TUPLE, std::tuple<ENTRY>>),
-                  "");
-
-    virtual void Process(typename YT::T_CONTAINER::type& container,
-                         typename YT::T_STREAM_TYPE& stream) override {
-      // TODO(dkorolev): Move this logic to the *RIGHT* place.
-      // TODO(max+dima): Ensure that this storage update can't break the actual state of the data.
-
-      typename YT::T_ABSTRACT_VISITABLE* av = dynamic_cast<typename YT::T_ABSTRACT_VISITABLE*>(entry.get());
-      if (av) {
-        av->accept(container);
-      } else {
-        // TODO(dkorolev): Talk to Max about API's policy on the stream containing an entry of unsupported type.
-        // Ex., have base entry type B, entry types X and Y, and finding an entry of type Z in the stream.
-        // Fail with an error message by default?
-        throw false;
-      }
-
-      static_cast<void>(stream);
-    }
-  };
-
-  // Sherlock stream listener call.
-  bool Entry(std::unique_ptr<ENTRY_BASE_TYPE>& entry, size_t index, size_t total) {
-    // The logic of this API implementation is:
-    // * Defer all API requests until the persistent part of the stream is fully replayed,
-    // * Allow all API requests after that.
-
-    mq_.EmplaceMessage(new MQMessageEntry(std::move(entry)));
-
-    // TODO(dkorolev): If that's the way to go, we should probably respond with HTTP 503 or 409 or 418?
-    //                 (And add a `/statusz` endpoint to monitor the status wrt ready / not yet ready.)
-
-    // TODO(dkorolev)+TODO(mzhurovich): What about the case of an empty stream?
-    // We should probably extend/enable Sherlock stream listener to report the `CaughtUp` event early,
-    // including the case where the initial stream is empty.
-
-    if (index + 1 == total) {
-      caught_up_ = true;
-    }
-
-    // This is primarily for unit testing purposes.
-    ++entries_seen_;
-
-    return true;
-  }
-
-  // Sherlock stream listener call.
-  void Terminate() {
-    // TODO(dkorolev): Should stop serving API requests.
-    // TODO(dkorolev): Should un-register HTTP endpoints, if they have been registered.
-  }
-
-  std::atomic_bool caught_up_;
-  std::atomic_size_t entries_seen_;
-
- private:
-  typename YT::T_MQ& mq_;
-};
-
-template <typename ENTRY>
-struct PolymorphicContainer<std::tuple<KeyEntry<ENTRY>>> {
-  typedef Container<KeyEntry<ENTRY>> type;
-};
-
-// TODO(dkorolev): This should be polymorphic- and variadic-friendly.
-template <typename ENTRY_BASE_TYPE, typename ENTRY>
-struct YodaImpl<ENTRY_BASE_TYPE, KeyEntry<ENTRY>> {
-  typedef YodaTypes<ENTRY_BASE_TYPE, std::tuple<KeyEntry<ENTRY>>> YT;
-
-  typedef ENTRY T_ENTRY;
-  typedef ENTRY_KEY_TYPE<T_ENTRY> T_KEY;
-
-  typedef std::function<void(const T_ENTRY&)> T_ENTRY_CALLBACK;
-  typedef std::function<void(const T_KEY&)> T_KEY_CALLBACK;
-  typedef std::function<void()> T_VOID_CALLBACK;
-
-  typedef KeyNotFoundException<T_ENTRY> T_KEY_NOT_FOUND_EXCEPTION;
-  typedef KeyAlreadyExistsException<T_ENTRY> T_KEY_ALREADY_EXISTS_EXCEPTION;
-  typedef EntryShouldExistException<T_ENTRY> T_ENTRY_SHOULD_EXIST_EXCEPTION;
-
-  YodaImpl() = delete;
-  explicit YodaImpl(typename YT::T_MQ& mq) : mq_(mq) {}
-
-  struct MQMessageGet : YT::T_MQ_MESSAGE {
-    const T_KEY key;
-    std::promise<T_ENTRY> pr;
-    T_ENTRY_CALLBACK on_success;
-    T_KEY_CALLBACK on_failure;
-
-    explicit MQMessageGet(const T_KEY& key, std::promise<T_ENTRY>&& pr) : key(key), pr(std::move(pr)) {}
-    explicit MQMessageGet(const T_KEY& key, T_ENTRY_CALLBACK on_success, T_KEY_CALLBACK on_failure)
-        : key(key), on_success(on_success), on_failure(on_failure) {}
-    virtual void Process(typename YT::T_CONTAINER::type& container, typename YT::T_STREAM_TYPE&) override {
-      const auto cit = container.data.find(key);
-      if (cit != container.data.end()) {
-        // The entry has been found.
-        if (on_success) {
-          // Callback semantics.
-          on_success(cit->second);
-        } else {
-          // Promise semantics.
-          pr.set_value(cit->second);
-        }
-      } else {
-        // The entry has not been found.
-        if (on_failure) {
-          // Callback semantics.
-          on_failure(key);
-        } else {
-          // Promise semantics.
-          SetPromiseToNullEntryOrThrow<T_KEY,
-                                       T_ENTRY,
-                                       T_KEY_NOT_FOUND_EXCEPTION,
-                                       false  // Was `T_POLICY::allow_nonthrowing_get>::DoIt(key, pr);`
-                                       >::DoIt(key, pr);
-        }
-      }
-    }
-  };
-
-  struct MQMessageAdd : YT::T_MQ_MESSAGE {
-    const T_ENTRY e;
-    std::promise<void> pr;
-    T_VOID_CALLBACK on_success;
-    T_VOID_CALLBACK on_failure;
-
-    explicit MQMessageAdd(const T_ENTRY& e, std::promise<void>&& pr) : e(e), pr(std::move(pr)) {}
-    explicit MQMessageAdd(const T_ENTRY& e, T_VOID_CALLBACK on_success, T_VOID_CALLBACK on_failure)
-        : e(e), on_success(on_success), on_failure(on_failure) {}
-
-    // Important note: The entry added will eventually reach the storage via the stream.
-    // Thus, in theory, `MQMessageAdd::Process()` could be a no-op.
-    // This code still updates the storage, to have the API appear more lively to the user.
-    // Since the actual implementation of `Add` pushes the `MQMessageAdd` message before publishing
-    // an update to the stream, the final state will always be [eventually] consistent.
-    // The practical implication here is that an API `Get()` after an api `Add()` may and will return data,
-    // that might not yet have reached the storage, and thus relying on the fact that an API `Get()` call
-    // reflects updated data is not reliable from the point of data synchronization.
-    virtual void Process(typename YT::T_CONTAINER::type& container,
-                         typename YT::T_STREAM_TYPE& stream) override {
-      const bool key_exists = static_cast<bool>(container.data.count(GetKey(e)));
-      if (key_exists) {
-        if (on_failure) {  // Callback function defined.
-          on_failure();
-        } else {  // Throw.
-          pr.set_exception(std::make_exception_ptr(T_KEY_ALREADY_EXISTS_EXCEPTION(e)));
-        }
-      } else {
-        container.data[GetKey(e)] = e;
-        stream.Publish(e);
-        if (on_success) {
-          on_success();
-        } else {
-          pr.set_value();
-        }
-      }
-    }
-  };
-
-  std::future<T_ENTRY> AsyncGet(const T_KEY& key) {
-    std::promise<T_ENTRY> pr;
-    std::future<T_ENTRY> future = pr.get_future();
-    mq_.EmplaceMessage(new MQMessageGet(key, std::move(pr)));
-    return future;
-  }
-
-  void AsyncGet(const T_KEY& key, T_ENTRY_CALLBACK on_success, T_KEY_CALLBACK on_failure) {
-    mq_.EmplaceMessage(new MQMessageGet(key, on_success, on_failure));
-  }
-
-  T_ENTRY Get(const T_KEY& key) { return AsyncGet(std::forward<const T_KEY>(key)).get(); }
-
-  std::future<void> AsyncAdd(const T_ENTRY& entry) {
-    std::promise<void> pr;
-    std::future<void> future = pr.get_future();
-
-    mq_.EmplaceMessage(new MQMessageAdd(entry, std::move(pr)));
-    return future;
-  }
-
-  void AsyncAdd(const T_ENTRY& entry,
-                T_VOID_CALLBACK on_success,
-                T_VOID_CALLBACK on_failure = [](const T_KEY&) {}) {
-    mq_.EmplaceMessage(new MQMessageAdd(entry, on_success, on_failure));
-  }
-
-  void Add(const T_ENTRY& entry) { AsyncAdd(entry).get(); }
-
- private:
-  typename YT::T_MQ& mq_;
-};
-
-template <typename ENTRY_BASE_TYPE, typename ENTRY>
-struct CombinedYodaImpls<ENTRY_BASE_TYPE, std::tuple<KeyEntry<ENTRY>>>
-    : YodaImpl<ENTRY_BASE_TYPE, KeyEntry<ENTRY>> {
-  typedef YodaTypes<ENTRY_BASE_TYPE, std::tuple<KeyEntry<ENTRY>>> YT;
-
-  static_assert(std::is_base_of<ENTRY_BASE_TYPE, ENTRY>::value,
-                "The first template parameter for `yoda::API` should be the base class for entry types.");
-
-  /*
-  // TODO(dkorolev): Give it another big thought. Right now `ENTRY_BASE_TYPE` is *NOT* visitable,
-  // otherwise the `visitor` implementation breaks due to ambiguous function resolution.
-  static_assert(
-      std::is_base_of<bricks::metaprogramming::abstract_visitable<typename YT::T_VISITABLE_TYPES_AS_TUPLE>,
-                      ENTRY_BASE_TYPE>::value,
-      "Entry type(s) for `yoda::API` should be `abstract_visitable<>` by the types used"
-      " in the API being instantiated. Please make your entry type(s) derive from `visitable<>`.");
-  */
-
-  CombinedYodaImpls() = delete;
-  explicit CombinedYodaImpls(typename YT::T_MQ& mq) : YodaImpl<ENTRY_BASE_TYPE, KeyEntry<ENTRY>>(mq) {}
-};
 
 // TODO(dkorolev): Base classes list should eventually go via `bricks::metaprogramming::combine<>`.
 template <typename ENTRY_BASE_TYPE, typename SUPPORTED_TYPES_AS_TUPLE>
